# swagger.yml
openapi: 3.0.3
info:
  title: "NUPEVID API"
  description: "API do sistema NUPEVID para gerenciamento de vítimas, atendimentos e medidas protetivas"
  version: "1.0.0"
  contact:
    name: "Equipe NUPEVID"
servers:
  - url: "/api/v1"
    description: "Servidor Principal"
...
security:
  - bearerAuth: []
  - api_key: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Autenticação via Token JWT obtido no endpoint /auth/login"
    api_key:
      type: apiKey
      in: header
      name: api_key
      description: "Chave de API configurada nas variáveis de ambiente"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Tipo de erro"
          example: "Bad Request"
        message:
          type: string
          description: "Mensagem de erro descritiva"
          example: "Validation error: field X is required"
        status_code:
          type: integer
          format: int32
          description: "Código de status HTTP"
          example: 400
...
    PhoneType:
      type: string
      enum:
        - "Mobile"
        - "Residential"
        - "Work"
      description: "Tipo de telefone"
...
    HasChildren:
      type: boolean
      description: "Indicador de filhos da vítima"
...
    EducationLevel:
      type: string
      enum:
        - "Elementary"
        - "High School"
        - "College"
        - "Postgraduate"
        - "Illiterate"
        - "Semi-illiterate"
        - "Master"
        - "Doctorate"
      description: "Nível de escolaridade"
...
    VictimPhone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "(69) 99999-8888"
        phone_type:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PhoneType"
    PhoneData:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "(69) 99999-8888"
        phone_type:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PhoneType"
    AddressType:
      type: string
      enum:
        - "Residential"
        - "Work"
        - "Correspondence"
        - "Commercial"
        - "Institutional"
        - "Temporary"
        - "Other"
      description: "Tipo de endereco da vitima"
    Victim:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        cpf:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        city_id:
          type: string
          format: uuid
        education_level:
          allOf:
            - $ref: "#/components/schemas/EducationLevel"
          nullable: true
        occupation:
          type: string
          nullable: true
          example: "Professora"
        has_children:
          $ref: "#/components/schemas/HasChildren"
        children_count:
          type: integer
          nullable: true
          example: 2
        is_pregnant:
          type: boolean
          nullable: true
        has_special_needs:
          type: boolean
        special_needs_type:
          type: array
          items:
            type: string
          nullable: true
        uses_alcohol:
          type: boolean
        uses_drugs:
          type: boolean
        has_psychiatric_issues:
          type: boolean
        psychiatric_issues_type:
          type: array
          items:
            type: string
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/VictimPhone"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/VictimAddress"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
...
    VictimAddress:
      type: object
      required:
        - address_type
      properties:
        id:
          type: string
          format: uuid
        street:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city_id:
          type: string
          format: uuid
          nullable: true
        zip_code:
          type: string
          nullable: true
        complement:
          type: string
          nullable: true
        address_type:
          $ref: "#/components/schemas/AddressType"
    AddressData:
      type: object
      required:
        - city_id
        - address_type
      properties:
        street:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city_id:
          type: string
          format: uuid
        zip_code:
          type: string
          nullable: true
        complement:
          type: string
          nullable: true
        address_type:
          $ref: "#/components/schemas/AddressType"
    CreateVictimRequest:
      type: object
      required:
        - full_name
        - uses_alcohol
        - uses_drugs
      properties:
        full_name:
          type: string
        cpf:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        city_id:
          type: string
          format: uuid
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneData"
          nullable: true
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/AddressData"
          nullable: true
        education_level:
          allOf:
            - $ref: "#/components/schemas/EducationLevel"
          nullable: true
        occupation:
          type: string
          nullable: true
        children_count:
          type: integer
          nullable: true
          description: "Número de filhos. Se informado, has_children será automaticamente definido como true"
        is_pregnant:
          type: boolean
          nullable: true
        special_needs_type:
          type: array
          items:
            type: string
          nullable: true
          description: "Tipos de necessidades especiais. Se informado com itens, has_special_needs será automaticamente definido como true"
        uses_alcohol:
          type: boolean
        uses_drugs:
          type: boolean
        psychiatric_issues_type:
          type: array
          items:
            type: string
          nullable: true
          description: "Tipos de problemas psiquiátricos. Se informado com itens, has_psychiatric_issues será automaticamente definido como true"
    UpdateVictimRequest:
      type: object
      required:
        - full_name
        - uses_alcohol
        - uses_drugs
      properties:
        full_name:
          type: string
        cpf:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        city_id:
          type: string
          format: uuid
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneData"
          nullable: true
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/AddressData"
          nullable: true
        education_level:
          allOf:
            - $ref: "#/components/schemas/EducationLevel"
          nullable: true
        occupation:
          type: string
          nullable: true
        children_count:
          type: integer
          nullable: true
          description: "Número de filhos. Se informado, has_children será automaticamente definido como true"
        is_pregnant:
          type: boolean
          nullable: true
        special_needs_type:
          type: array
          items:
            type: string
          nullable: true
          description: "Tipos de necessidades especiais. Se informado com itens, has_special_needs será automaticamente definido como true"
        uses_alcohol:
          type: boolean
        uses_drugs:
          type: boolean
        psychiatric_issues_type:
          type: array
          items:
            type: string
          nullable: true
          description: "Tipos de problemas psiquiátricos. Se informado com itens, has_psychiatric_issues será automaticamente definido como true"

    # --- Offender Schemas ---
...
paths:
...
  /victims:
    get:
      tags:
        - "Victims"
      summary: "Listar todas as vítimas"
      description: "Retorna lista de vítimas com filtros baseados nas permissões do usuário"
      operationId: "getAllVictims"
      security:
        - bearerAuth: []
          api_key: []
      parameters:
        - name: "page"
          in: "query"
          required: false
          schema:
            type: integer
            example: 1
          description: "Número da página (1-based)"
        - name: "page_size"
          in: "query"
          required: false
          schema:
            type: integer
            example: 10
          description: "Quantidade de itens por página (max 200)"
      responses:
        "200":
          description: "Lista de vítimas retornada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Victim"
                  page:
                    type: integer
                    example: 1
                  page_size:
                    type: integer
                    example: 10
                  total_items:
                    type: integer
                    example: 100
                  total_pages:
                    type: integer
                    example: 10
              example:
                message: "Operation successful"
                status: 200
                data:
                  - id: "770e8400-e29b-41d4-a716-446655440000"
                    full_name: "Ana Paula dos Santos"
                    cpf: "529.982.247-25"
                    birth_date: "1990-05-15"
                    city_id: "660e8400-e29b-41d4-a716-446655440000"
                    created_at: "2025-01-15T10:30:00Z"
                    updated_at: "2025-01-15T10:30:00Z"
                    is_deleted: false
                    education_level: "High School"
                    occupation: "Professora"
                    has_children: true
                    children_count: 2
                    is_pregnant: false
                    has_special_needs: false
                    special_needs_type: null
                    uses_alcohol: false
                    uses_drugs: false
                    has_psychiatric_issues: false
                    psychiatric_issues_type: null
                    phones:
                      - id: "880e8400-e29b-41d4-a716-446655440000"
                        phone: "(69) 99999-8888"
                        phone_type: "Mobile"
                    addresses:
                      - id: "990e8400-e29b-41d4-a716-446655440000"
                        street: "Rua das Flores"
                        number: "123"
                        district: "Centro"
                        city_id: "660e8400-e29b-41d4-a716-446655440000"
                        zip_code: "76801-000"
                        complement: null
                        address_type: "Residential"
                page: 1
                page_size: 10
                total_items: 1
                total_pages: 1
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para listar vítimas"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403

    post:
      tags:
        - "Victims"
      summary: "Criar nova vítima"
      description: "Cria uma nova vítima no sistema"
      operationId: "createVictim"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVictimRequest"
      responses:
        "201":
          description: "Vítima criada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  data:
                    $ref: "#/components/schemas/Victim"
              example:
                message: "Operation successful"
                status: 201
                data:
                  id: "770e8400-e29b-41d4-a716-446655440000"
                  full_name: "Ana Paula dos Santos"
                  cpf: "529.982.247-25"
                  birth_date: "1990-05-15"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-01-15T10:30:00Z"
                  is_deleted: false
                  education_level: "High School"
                  occupation: "Professora"
                  has_children: true
                  children_count: 2
                  is_pregnant: false
                  has_special_needs: false
                  special_needs_type: null
                  uses_alcohol: false
                  uses_drugs: false
                  has_psychiatric_issues: false
                  psychiatric_issues_type: null
                  phones: []
                  addresses: []
        "400":
          description: "Dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: missing field `full_name`"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para criar vítima nesta cidade"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
...
  /victims/{id}:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: string
          format: uuid
        description: "ID da vítima"

    get:
      tags:
        - "Victims"
      summary: "Obter vítima por ID"
      description: "Retorna detalhes de uma vítima específica"
      operationId: "getVictimById"
      security:
        - bearerAuth: []
          api_key: []
      responses:
        "200":
          description: "Vítima encontrada"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/Victim"
              example:
                message: "Operation successful"
                status: 200
                data:
                  id: "770e8400-e29b-41d4-a716-446655440000"
                  full_name: "Ana Paula dos Santos"
                  cpf: "529.982.247-25"
                  birth_date: "1990-05-15"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-01-15T10:30:00Z"
                  is_deleted: false
                  education_level: "High School"
                  occupation: "Professora"
                  has_children: true
                  children_count: 2
                  is_pregnant: false
                  has_special_needs: false
                  special_needs_type: null
                  uses_alcohol: false
                  uses_drugs: false
                  has_psychiatric_issues: false
                  psychiatric_issues_type: null
                  phones: []
                  addresses: []
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para visualizar esta vítima"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Vítima não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: victim not found"
                status_code: 404

    put:
      tags:
        - "Victims"
      summary: "Atualizar vítima"
      description: "Atualiza informações de uma vítima existente"
      operationId: "updateVictim"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVictimRequest"
      responses:
        "200":
          description: "Vítima atualizada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/Victim"
        "400":
          description: "Dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: validation error"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para atualizar esta vítima"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Vítima não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: victim not found"
                status_code: 404

    delete:
      tags:
        - "Victims"
      summary: "Deletar vítima (soft delete)"
      description: "Realiza soft delete de uma vítima"
      operationId: "deleteVictim"
      security:
        - bearerAuth: []
          api_key: []
      responses:
        "200":
          description: "Vítima deletada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/Victim"
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para deletar esta vítima"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Vítima não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: victim not found"
                status_code: 404
...